+++ test/system/projects/samples_test.rb
@@ -325,308 +325,6 @@ module Projects
       ### VERIFY END ###
     end
 
-    test 'transfer dialog sample listing' do
-      ### SETUP START ###
-      samples = @project.samples.pluck(:puid, :name)
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.transfer_button')
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#list_selections') do
-        samples.each do |sample|
-          assert_text sample[0]
-          assert_text sample[1]
-        end
-      end
-      ### VERIFY END ###
-    end
-
-    test 'transfer dialog with plural description' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.transfer_button')
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#dialog') do
-        assert_text I18n.t('projects.samples.transfers.dialog.description.plural').gsub!('COUNT_PLACEHOLDER',
-                                                                                         '3')
-      end
-      ### VERIFY END ###
-    end
-
-    test 'transfer dialog with singular description' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      within '#samples-table table tbody' do
-        all('input[type="checkbox"]')[0].click
-      end
-      click_link I18n.t('projects.samples.index.transfer_button')
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#dialog') do
-        assert_text I18n.t('projects.samples.transfers.dialog.description.singular')
-      end
-      ### VERIFY END ###
-    end
-
-    test 'transfer samples' do
-      ### SETUP START ###
-      samples = @project.samples.pluck(:puid, :name)
-      # show destination project has 20 samples prior to transfer
-      visit namespace_project_samples_url(@namespace, @project2)
-      assert_text strip_tags(I18n.t(:'viral.pagy.limit_component.summary', from: 1, to: 20, count: 20,
-                                                                           locale: @user.locale))
-      # originating project has 3 samples prior to transfer
-      visit namespace_project_samples_url(@namespace, @project)
-      assert_text strip_tags(I18n.t(:'viral.pagy.limit_component.summary', from: 1, to: 3, count: 3,
-                                                                           locale: @user.locale))
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      # select all 3 samples
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.transfer_button')
-      within('#dialog') do
-        within('#list_selections') do
-          samples.each do |sample|
-            assert_text sample[0]
-            assert_text sample[1]
-          end
-        end
-        # select destination project
-        find('input#select2-input').click
-        find("button[data-viral--select2-primary-param='#{@project2.full_path}']").click
-        click_on I18n.t('projects.samples.transfers.dialog.submit_button')
-      end
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      # flash msg
-      assert_text I18n.t('projects.samples.transfers.create.success')
-      # originating project no longer has samples
-      assert_text I18n.t('projects.samples.index.no_samples')
-
-      # destination project received transferred samples
-      visit namespace_project_samples_url(@namespace, @project2)
-      within '#samples-table table tbody' do
-        samples.each do |sample|
-          assert_text sample[0]
-          assert_text sample[1]
-        end
-      end
-      assert_text strip_tags(I18n.t(:'viral.pagy.limit_component.summary', from: 1, to: 20, count: 23,
-                                                                           locale: @user.locale))
-      ### VERIFY END ###
-    end
-
-    test 'should not transfer samples with session storage cleared' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project)
-      assert_text strip_tags(I18n.t(:'viral.pagy.limit_component.summary', from: 1, to: 3, count: 3,
-                                                                           locale: @user.locale))
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      # select samples
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      # clear localstorage
-      Capybara.execute_script 'sessionStorage.clear()'
-      # launch transfer dialog
-      click_link I18n.t('projects.samples.index.transfer_button')
-      within('#dialog') do
-        find('input#select2-input').click
-        find("button[data-viral--select2-primary-param='#{@project2.full_path}']").click
-        click_on I18n.t('projects.samples.transfers.dialog.submit_button')
-        ### ACTIONS END ###
-
-        ### VERIFY START ###
-        # samples listing should no longer appear in dialog
-        assert_no_selector '#list_selections'
-        # error msg displayed in dialog
-        assert_text I18n.t('projects.samples.transfers.create.no_samples_transferred_error')
-      end
-      ### VERIFY END ###
-    end
-
-    test 'transfer samples with and without same name in destination project' do
-      # only samples without a matching name to samples in destination project will transfer
-
-      ### SETUP START ###
-      samples = @project.samples.pluck(:puid, :name)
-      namespace = groups(:subgroup1)
-      project25 = projects(:project25)
-
-      # verify only 2 samples exist in destination project
-      visit namespace_project_samples_url(namespace, project25)
-      assert_text strip_tags(I18n.t(:'viral.pagy.limit_component.summary', from: 1, to: 2, count: 2,
-                                                                           locale: @user.locale))
-      # 3 samples in originating project
-      visit namespace_project_samples_url(@namespace, @project)
-      assert_text strip_tags(I18n.t(:'viral.pagy.limit_component.summary', from: 1, to: 3, count: 3,
-                                                                           locale: @user.locale))
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.transfer_button')
-      within('#dialog') do
-        within('#list_selections') do
-          samples.each do |sample|
-            assert_text sample[0]
-            assert_text sample[1]
-          end
-        end
-        find('input#select2-input').click
-        find("button[data-viral--select2-primary-param='#{project25.full_path}']").click
-        click_on I18n.t('projects.samples.transfers.dialog.submit_button')
-      end
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#dialog') do
-        # error messages in dialog
-        assert_text I18n.t('projects.samples.transfers.create.error')
-        # colon is removed from translation in UI
-        assert_text I18n.t('services.samples.transfer.sample_exists', sample_puid: @sample30.puid,
-                                                                      sample_name: @sample30.name).gsub(':', '')
-      end
-
-      # verify sample1 and 2 transferred, sample 30 did not
-      assert_text strip_tags(I18n.t(:'viral.pagy.limit_component.summary.one', count: 1, locale: @user.locale))
-      assert_no_selector "tr[id='#{@sample1.id}']"
-      assert_no_selector "tr[id='#{@sample2.id}']"
-      assert_selector "tr[id='#{@sample30.id}']"
-
-      # destination project
-      visit namespace_project_samples_url(namespace, project25)
-      assert_text strip_tags(I18n.t(:'viral.pagy.limit_component.summary', from: 1, to: 4, count: 4,
-                                                                           locale: @user.locale))
-      assert_selector "tr[id='#{@sample1.id}']"
-      assert_selector "tr[id='#{@sample2.id}']"
-      assert_no_selector "tr[id='#{@sample30.id}']"
-      ### VERIFY END ###
-    end
-
-    test 'sample transfer project listing should be empty for maintainer if no other projects in hierarchy' do
-      ### SETUP START ###
-      login_as users(:user28)
-      namespace = groups(:group_hotel)
-      project = projects(:projectHotel)
-      visit namespace_project_samples_url(namespace, project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.transfer_button')
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#dialog') do
-        # no available destination projects
-        assert_selector "input[placeholder='#{I18n.t('projects.samples.transfers.dialog.no_available_projects')}']"
-      end
-      ### VERIFY END ###
-    end
-
-    test 'updating sample selection during transfer samples' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project2)
-
-      # verify no samples currently selected in destination project
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 20"
-        assert_selector 'strong[data-selection-target="selected"]', text: '0'
-      end
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      # select 1 sample to transfer
-      within '#samples-table table tbody' do
-        all('input[type="checkbox"]')[0].click
-      end
-
-      # verify 1 sample selected in originating project
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 3"
-        assert_selector 'strong[data-selection-target="selected"]', text: '1'
-      end
-
-      # transfer sample
-      click_link I18n.t('projects.samples.index.transfer_button')
-      within('#dialog') do
-        within('#list_selections') do
-          assert_text @sample1.name
-          assert_text @sample1.puid
-        end
-        find('input#select2-input').click
-        click_button @project2.puid
-        # find("button[data-viral--select2-primary-param='#{@project2.full_path}']").click
-        click_on I18n.t('projects.samples.transfers.dialog.submit_button')
-      end
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      # verify no samples selected anymore
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 2"
-        assert_selector 'strong[data-selection-target="selected"]', text: '0'
-      end
-
-      # verify destination project still has no selected samples and one additional sample
-      visit namespace_project_samples_url(@namespace, @project2)
-      assert_selector 'h1', text: 'Samples'
-      assert_selector '#samples-table'
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 21"
-        assert_selector 'strong[data-selection-target="selected"]', text: '0'
-      end
-
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 21"
-        assert_selector 'strong[data-selection-target="selected"]', text: '21'
-      end
-      ### VERIFY END
-    end
-
-    test 'empty state of transfer sample project selection' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      # select samples
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-
-      # launch dialog
-      click_link I18n.t('projects.samples.index.transfer_button')
-      within('#dialog') do
-        # fill destination input
-        find('input#select2-input').fill_in with: 'invalid project name or puid'
-        ### ACTIONS END ###
-
-        ### VERIFY START ###
-        assert_text I18n.t('projects.samples.transfers.dialog.empty_state')
-        ### VERIFY END ###
-      end
-    end
-
     test 'can search the list of samples by name' do
       ### SETUP START ###
       visit namespace_project_samples_url(@namespace, @project)
@@ -1379,288 +1077,6 @@ module Projects
         end
         ### VERIFY END ###
       end
-    end
-
-    test 'clone dialog sample listing' do
-      ### SETUP START ###
-      samples = @project.samples.pluck(:puid, :name)
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.clone_button')
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#list_selections') do
-        samples.each do |sample|
-          assert_text sample[0]
-          assert_text sample[1]
-        end
-      end
-      ### VERIFY END ###
-    end
-
-    test 'singular clone dialog description' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      within '#samples-table table tbody' do
-        all('input[type="checkbox"]')[0].click
-      end
-      click_link I18n.t('projects.samples.index.clone_button')
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#dialog') do
-        assert_text I18n.t('projects.samples.clones.dialog.description.singular')
-      end
-      ### VERIFY END ###
-    end
-
-    test 'plural clone dialog description' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.clone_button')
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#dialog') do
-        assert_text I18n.t(
-          'projects.samples.clones.dialog.description.plural'
-        ).gsub! 'COUNT_PLACEHOLDER', '3'
-      end
-      ### VERIFY END ###
-    end
-
-    test 'should clone samples' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project2)
-      # verify samples 1 and 2 do not exist in project2
-      within('#samples-table table tbody') do
-        assert_no_text @sample1.name
-        assert_no_text @sample2.name
-      end
-
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      # select samples 1 and 2 for cloning
-      within '#samples-table table tbody' do
-        find("input#sample_#{@sample1.id}").click
-        find("input#sample_#{@sample2.id}").click
-      end
-      click_link I18n.t('projects.samples.index.clone_button')
-      within('#dialog') do
-        within('#list_selections') do
-          assert_text @sample1.puid
-          assert_text @sample1.name
-          assert_text @sample2.puid
-          assert_text @sample2.name
-        end
-        find('input#select2-input').click
-        find("button[data-viral--select2-primary-param='#{@project2.full_path}']").click
-        click_on I18n.t('projects.samples.clones.dialog.submit_button')
-      end
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      # flash msg
-      assert_text I18n.t('projects.samples.clones.create.success')
-      # samples still exist within samples table of originating project
-      within('#samples-table table tbody') do
-        assert_text @sample1.name
-        assert_text @sample2.name
-      end
-
-      # samples now exist in project2 samples table
-      visit namespace_project_samples_url(@namespace, @project2)
-      within('#samples-table table tbody') do
-        assert_text @sample1.name
-        assert_text @sample2.name
-      end
-      ### VERIFY END ###
-    end
-
-    test 'should not clone samples with session storage cleared' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      # clear localstorage
-      Capybara.execute_script 'sessionStorage.clear()'
-      click_link I18n.t('projects.samples.index.clone_button')
-      within('#dialog') do
-        find('input#select2-input').click
-        find("button[data-viral--select2-primary-param='#{@project2.full_path}']").click
-        click_on I18n.t('projects.samples.clones.dialog.submit_button')
-        ### ACTIONS END ###
-
-        ### VERIFY START ###
-        # sample listing should not be in error dialog
-        assert_no_selector '#list_selections'
-        # error msg
-        assert_text I18n.t('projects.samples.clones.create.no_samples_cloned_error')
-        assert_text I18n.t('services.samples.clone.empty_sample_ids')
-        ### VERIFY END ###
-      end
-    end
-
-    test 'should not clone some samples' do
-      ### SETUP START ###
-      namespace = groups(:subgroup1)
-      project25 = projects(:project25)
-      samples = @project.samples.pluck(:puid, :name)
-      visit namespace_project_samples_url(namespace, project25)
-      # sample30's name already exists in project25 samples table, samples1 and 2 do not
-      within('#samples-table table tbody') do
-        assert_no_text @sample1.name
-        assert_no_text @sample2.name
-        assert_text @sample30.name
-      end
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.clone_button')
-      within('#dialog') do
-        within('#list_selections') do
-          samples.each do |sample|
-            assert_text sample[0]
-            assert_text sample[1]
-          end
-        end
-        find('input#select2-input').click
-        find("button[data-viral--select2-primary-param='#{project25.full_path}']").click
-        click_on I18n.t('projects.samples.clones.dialog.submit_button')
-
-        ### ACTIONS END ###
-
-        ### VERIFY START ###
-        # errors that a sample with the same name as sample30 already exists in project25
-        assert_text I18n.t('projects.samples.clones.create.error')
-        assert_text I18n.t('services.samples.clone.sample_exists', sample_puid: @sample30.puid,
-                                                                   sample_name: @sample30.name).gsub(':', '')
-        click_on I18n.t('projects.samples.shared.errors.ok_button')
-      end
-
-      visit namespace_project_samples_url(namespace, project25)
-      # samples 1 and 2 still successfully clone
-      within('#samples-table table tbody') do
-        assert_text @sample1.name
-        assert_text @sample2.name
-      end
-      ### VERIFY END ###
-    end
-
-    test 'empty state of destination project selection for sample cloning' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ####
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.clone_button')
-      within('#dialog') do
-        find('input#select2-input').fill_in with: 'invalid project name or puid'
-        ### ACTIONS END ###
-
-        ### VERIFY START ###
-        assert_text I18n.t('projects.samples.clones.dialog.empty_state')
-        ### VERIFY END ###
-      end
-    end
-
-    test 'no available destination projects to clone samples' do
-      ### SETUP START ###
-      sign_in users(:jean_doe)
-      visit namespace_project_samples_url(namespaces_user_namespaces(:john_doe_namespace), projects(:john_doe_project2))
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-      click_link I18n.t('projects.samples.index.clone_button')
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      within('#dialog') do
-        assert "input[placeholder='#{I18n.t('projects.samples.clones.dialog.no_available_projects')}']"
-      end
-      ### VERIFY END ###
-    end
-
-    test 'updating sample selection during sample cloning' do
-      ### SETUP START ###
-      visit namespace_project_samples_url(@namespace, @project2)
-      # verify no samples currently selected in destination project
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 20"
-        assert_selector 'strong[data-selection-target="selected"]', text: '0'
-      end
-      visit namespace_project_samples_url(@namespace, @project)
-      ### SETUP END ###
-
-      ### ACTIONS START ###
-      # select 1 sample to clone
-      within '#samples-table table tbody' do
-        all('input[type="checkbox"]')[0].click
-      end
-
-      # verify 1 sample selected in originating project
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 3"
-        assert_selector 'strong[data-selection-target="selected"]', text: '1'
-      end
-
-      # clone sample
-      click_link I18n.t('projects.samples.index.clone_button')
-      within('#dialog') do
-        within('#list_selections') do
-          assert_text @sample1.name
-          assert_text @sample1.puid
-        end
-        find('input#select2-input').click
-        find("button[data-viral--select2-primary-param='#{@project2.full_path}']", wait: 1).click
-        click_on I18n.t('projects.samples.clones.dialog.submit_button')
-      end
-      ### ACTIONS END ###
-
-      ### VERIFY START ###
-      sleep 1
-      # flash msg
-      assert_text I18n.t('projects.samples.clones.create.success')
-      # verify no samples selected anymore
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 3"
-        assert_selector 'strong[data-selection-target="selected"]', text: '0'
-      end
-
-      # verify destination project still has no selected samples and one additional sample
-      visit namespace_project_samples_url(@namespace, @project2)
-
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 21"
-        assert_selector 'strong[data-selection-target="selected"]', text: '0'
-      end
-
-      click_button I18n.t(:'projects.samples.index.select_all_button')
-
-      within 'tfoot' do
-        assert_text "#{I18n.t('samples.table_component.counts.samples')}: 21"
-        assert_selector 'strong[data-selection-target="selected"]', text: '21'
-      end
-      ### VERIFY END
     end
 
     test 'selecting / deselecting all samples' do
